refactor: Moved ConsoleColors.java to parade.utils pkg
